Moved signature encoding code into a separate module.
