+++ libcrux-ml-dsa/src/encoding.rs
@@ -1,6 +1,7 @@
 pub(crate) mod commitment;
 pub(crate) mod error;
 pub(crate) mod gamma1;
+pub(crate) mod signature;
 pub(crate) mod signing_key;
 pub(crate) mod t0;
 pub(crate) mod t1;
